Create a private wrapper checkbox component
This allows us to split off the group logic from the core checkbox logic, which makes the component easier to maintain.
