+++ addon/components/au-checkbox-group.hbs
@@ -7,8 +7,7 @@
   {{yield
     (hash
       Checkbox=(component
-        "au-checkbox"
-        inGroup=true
+        this.GroupCheckbox
         groupValue=@value
         disabled=@disabled
         onChangeGroup=@onChange
