+++ addon/components/au-checkbox-group/checkbox.hbs
@@ -1,0 +1,7 @@
+<AuCheckbox
+  @checked={{this.isCheckedInGroup}}
+  @disabled={{@disabled}}
+  @name={{@name}}
+  @onChange={{this.onChange}}
+  ...attributes
+>{{yield}}</AuCheckbox>
